+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.1.3] - 2024-10-11
+### Added
+- Added "GridData Scriptable Object" for grid data
+  
+### Changed
+- Change class name "GridBase" -> "TileBase"
+- Change class name "GridNode" -> "TileNode"
+- Change class name "GridDisplay" -> "TileDisplay"
+- Split class "GridSpawner" -> "TileSpawner", "GridManager"
+
+
 ## [0.1.2] - 2024-10-09
 ### Added
 - Support for multiple grids